Numerous issues discovered during code review.
Amazing how many things are still technically wrong.  Probably should
have done templates.

+++ src/Integer64RealMap.F90
@@ -12,9 +12,9 @@ module gFTL_Integer64RealMap
 
 #else
 
-  use, intrinsic :: iso_fortran_env, only: INT32
+  use, intrinsic :: iso_fortran_env, only: INT64
  
-#  define _key type(integer(kind=INT32))
+#  define _key type(integer(kind=INT64))
 #  define _value type(real)
 #  define _map Integer64RealMap
 #  define _iterator Integer64RealMapIterator
